requirements.txt
streamlit
openai>=1.0
langchain
langchain-community
faiss-cpu
pypdf
python-dotenv
tiktoken
fastembed
